Nextcloud: Build custom nextcloud image
The vanilla Nextcloud upstream image does alot of initialization
that is not wanted for automatic installation of new nodes in an
existing Nextcloud cluster. This image will work better for SUNET
Drive although it will also mean that we will have to bootstrap
a new Nextcloud cluster independently of this image.